+++ libcnb/src/toml/document.ts
@@ -1,10 +1,10 @@
-import type { TomlTable }     from './table.js'
+import type { TomlTable }     from './readers/index.js'
 
 import { parse }              from '@iarna/toml'
 import { stringify }          from '@iarna/toml'
 
 import { InvalidConfigError } from '../errors/index.js'
-import { asTomlTable }        from './table.js'
+import { asTomlTable }        from './readers/index.js'
 
 export const parseTomlTable = (content: string, path: string): TomlTable => {
   try {
