+++ libcnb/src/toml/errors/invalid-value.ts
@@ -1,0 +1,4 @@
+import { InvalidConfigError } from '../../errors/index.js'
+
+export const invalidTomlValue = (path: string, expected: string): InvalidConfigError =>
+  new InvalidConfigError(`${path} must be ${expected}`)
