+++ libcnb/src/toml/file.ts
@@ -1,4 +1,4 @@
-import type { TomlTable }     from './table.js'
+import type { TomlTable }     from './readers/index.js'
 
 import { readFile }           from 'node:fs/promises'
 import { writeFile }          from 'node:fs/promises'
