+++ libcnb/src/toml/index.ts
@@ -1,3 +1,4 @@
 export * from './file.js'
 export * from './document.js'
-export * from './table.js'
+export * from './errors/index.js'
+export * from './readers/index.js'
